add filter for get examples, easy first and exclude account examples

+++ run_spider2.py
@@ -51,7 +51,7 @@ def config() -> argparse.Namespace:
     parser = argparse.ArgumentParser(description="Run end-to-end evaluation on the benchmark")
 
     # environment config
-    parser.add_argument("--path_to_vm", type=str)
+    parser.add_argument("--path_to_vm", type=str, help="path to the VM executable .vmx file, if None, automatically find the VM in vm_data/ folder")
     parser.add_argument("--snapshot_name", type=str, default="spider2.0", help="Snapshot name to use (overwrite snapshot in each example config)")
     parser.add_argument("--headless", action="store_true", help="Run in headless machine")
     parser.add_argument(
@@ -201,12 +201,14 @@ def get_result_dir(args):
     return os.path.join(args.result_dir, result_dir)
 
 
-def get_examples(args, result_dir) -> List[Dict[str, str]]:
+def get_examples(args, result_dir, easy_first: bool = True, exclude_account: bool = True) -> List[Dict[str, str]]:
     """ Get [Filter] the list of example dict for the current experiment.
     # Filter method:
     - args.from_scratch (bool): if True, ignore existing results under the result directory, otherwise,
         only include examples that do not have `result.txt` file under the result directory
     - args.domains (List[str]): if not contain "all", only include examples under the specified domain/tool
+    - easy_first (bool): if True, include examples that are easy to run first (smaller action_number)
+    - exclude_account (bool): if True, exclude examples that are related to real accounts
     
     # The returned dict for each example in the List containing:
         - id: example id
@@ -236,6 +238,10 @@ def get_examples(args, result_dir) -> List[Dict[str, str]]:
                 example_result_dir = os.path.join(domain_result_dir, example_id)
                 result_file = os.path.join(example_result_dir, "result.txt")
                 if not args.from_scratch and os.path.exists(result_file) and file_not_empty(result_file): continue
+                data_config = os.path.join(example_dir, f"{example_id}.json")
+                data = json.load(open(data_config, 'r'))
+                if exclude_account and 'account' in data['tags']: continue
+
                 # remove the result directory if exists
                 shutil.rmtree(example_result_dir, ignore_errors=True)
                 os.makedirs(example_result_dir, exist_ok=True)
@@ -246,11 +252,13 @@ def get_examples(args, result_dir) -> List[Dict[str, str]]:
                 example = {
                     "id": example_id,
                     "domain": domain,
-                    "config": os.path.join(example_dir, f"{example_id}.json"),
-                    "result": example_result_dir
+                    "config": data_config,
+                    "result": example_result_dir,
+                    "action_number": data["action_number"]
                 }
                 examples_to_run.append(example)
     logger.info(f"Total examples to run: {len(examples_to_run)}")
+    if easy_first: sorted(examples_to_run, key=lambda x: x['action_number'])
     return examples_to_run
